Fix infinite loop for self-loops in edge graph
The PR review process caught an infinite loop for polygons containing duplicate sequential coordinates (excluding start and end points). This created a self loop in the edge graph that caused an infinite loop during traversal. One proposed solution was to check for duplicate points, ignore, and proceed. This violates the OGC SFA definition of a Simple Curve (the base of a LinearRing). Per 6.1.6.1 in v.1.2.1, "A Curve is simple if it does not pass through the same point twice with the possible exception of the two end points...A curve that is simple and closed is a Ring".

Keeping true to "fail early", this commit catches self-loops during creation of the edge and throws an IllegalShapeException. It complys with the OGC spec and avoids any further processing if an invalid ring is provided. One potential downside is the additional rigor it adds to loosely defined GeoJSON.